feat: 라우터 설치 및 높이조절 커스텀 훅으로 분리
