+++ src/App.tsx
@@ -1,11 +1,21 @@
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { RecoilRoot } from "recoil";
 import { BrowserView, MobileView } from "react-device-detect";
-import ChatRoomPage from "./pages/ChatRoomPage";
 
 import { ReactComponent as StatusBar } from "./assets/images/status_bar.svg";
 import { ReactComponent as HomeIndicator } from "./assets/images/home_indicator.svg";
 
+import MyProfilePage from "./pages/MyProfilePage";
+import FollowListPage from "./pages/FollowListPage";
+import ChatListPage from "./pages/ChatListPage";
+import ChatRoomPage from "./pages/ChatRoomPage";
+
+//hooks
+import useHandleHeight from "./hooks/useHandleHeight";
+
 function App() {
+  const height = useHandleHeight();
+
   return (
     <>
       <RecoilRoot>
@@ -13,7 +23,19 @@ function App() {
         <BrowserView>
           <div className="flex min-h-screen w-dvw flex-col items-center justify-center">
             <StatusBar />
-            <ChatRoomPage />
+            <div className="relative w-[375px] bg-white" style={{ height }}>
+              <Router>
+                <Routes>
+                  <Route path={"/"} element={<MyProfilePage />}></Route>
+                  <Route
+                    path={"/followlist"}
+                    element={<FollowListPage />}
+                  ></Route>
+                  <Route path={"/chatlist"} element={<ChatListPage />}></Route>
+                  <Route path={"/chatroom"} element={<ChatRoomPage />}></Route>
+                </Routes>
+              </Router>
+            </div>
             <HomeIndicator />
           </div>
         </BrowserView>
