+++ src/pages/ChatListPage.tsx
@@ -1,0 +1,9 @@
+const ChatListPage = () => {
+  return (
+    <>
+      <div>ChatListPage</div>
+    </>
+  );
+};
+
+export default ChatListPage;
