+++ src/pages/FollowListPage.tsx
@@ -1,0 +1,9 @@
+const FollowListPage = () => {
+  return (
+    <>
+      <div>FollowListPage</div>
+    </>
+  );
+};
+
+export default FollowListPage;
