+++ src/pages/MyProfilePage.tsx
@@ -1,0 +1,9 @@
+const MyProfilePage = () => {
+  return (
+    <>
+      <div>MyProfilePage</div>
+    </>
+  );
+};
+
+export default MyProfilePage;
